+++ src/rules/no-unused-modules.js
@@ -42,6 +42,7 @@ const VARIABLE_DECLARATION = 'VariableDeclaration'
 const FUNCTION_DECLARATION = 'FunctionDeclaration'
 const CLASS_DECLARATION = 'ClassDeclaration'
 const DEFAULT = 'default'
+const TYPE_ALIAS = 'TypeAlias'
 
 let preparationDone = false
 const importList = new Map()
@@ -463,7 +464,8 @@ module.exports = {
           if (declaration) {
             if (
               declaration.type === FUNCTION_DECLARATION ||
-              declaration.type === CLASS_DECLARATION
+              declaration.type === CLASS_DECLARATION ||
+              declaration.type === TYPE_ALIAS
             ) {
               newExportIdentifiers.add(declaration.id.name)
             }
@@ -788,7 +790,8 @@ module.exports = {
         if (node.declaration) {
           if (
             node.declaration.type === FUNCTION_DECLARATION ||
-            node.declaration.type === CLASS_DECLARATION
+            node.declaration.type === CLASS_DECLARATION ||
+            node.declaration.type === TYPE_ALIAS
           ) {
             checkUsage(node, node.declaration.id.name)
           }