+++ tests/files/no-unused-modules/flow-0.js
@@ -1,0 +1,1 @@
+import { type FooType } from './flow-2';