+++ tests/files/no-unused-modules/flow-1.js
@@ -1,0 +1,2 @@
+// @flow strict
+export type Bar = number;